Move vaccine type locale tests to test_locale.py
